chamados-manutencao-v2.0-completo-final/chamados-manutencao-distribuicao/sistema-chamados-railway/sistema-chamados-railway/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Dependências adicionais para produção
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
Pillow==10.1.0
python-dateutil==2.8.2
